Move the metrics stuff to its own class
